requirements.txt
opencv-python
pytesseract
pdf2image
pillow
numpy
text-unidecode